Allow for different implementations of StatementClient
