+++ presto-client/src/main/java/com/facebook/presto/client/StatementClient.java
@@ -13,449 +13,59 @@
  */
 package com.facebook.presto.client;
 
-import com.facebook.presto.client.OkHttpUtil.NullCallback;
 import com.facebook.presto.spi.type.TimeZoneKey;
-import com.google.common.base.Joiner;
-import com.google.common.base.Splitter;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Sets;
-import io.airlift.json.JsonCodec;
-import io.airlift.units.Duration;
-import okhttp3.Headers;
-import okhttp3.HttpUrl;
-import okhttp3.MediaType;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.RequestBody;
 
 import javax.annotation.Nullable;
-import javax.annotation.concurrent.ThreadSafe;
 
 import java.io.Closeable;
-import java.io.UnsupportedEncodingException;
-import java.net.URI;
-import java.net.URLDecoder;
-import java.net.URLEncoder;
-import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
 
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_ADDED_PREPARE;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_CATALOG;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_CLEAR_SESSION;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_CLEAR_TRANSACTION_ID;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_CLIENT_INFO;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_CLIENT_TAGS;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_DEALLOCATED_PREPARE;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_LANGUAGE;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_PREPARED_STATEMENT;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_SCHEMA;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_SESSION;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_SET_CATALOG;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_SET_SCHEMA;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_SET_SESSION;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_SOURCE;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_STARTED_TRANSACTION_ID;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_TIME_ZONE;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_TRANSACTION_ID;
-import static com.facebook.presto.client.PrestoHeaders.PRESTO_USER;
-import static com.google.common.base.MoreObjects.firstNonNull;
-import static com.google.common.base.Preconditions.checkState;
-import static com.google.common.net.HttpHeaders.USER_AGENT;
-import static io.airlift.json.JsonCodec.jsonCodec;
-import static java.lang.String.format;
-import static java.net.HttpURLConnection.HTTP_OK;
-import static java.net.HttpURLConnection.HTTP_UNAUTHORIZED;
-import static java.net.HttpURLConnection.HTTP_UNAVAILABLE;
-import static java.util.Objects.requireNonNull;
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
-
-@ThreadSafe
-public class StatementClient
-        implements Closeable
+public interface StatementClient
+        extends Closeable
 {
-    private static final MediaType MEDIA_TYPE_JSON = MediaType.parse("application/json; charset=utf-8");
-    private static final JsonCodec<QueryResults> QUERY_RESULTS_CODEC = jsonCodec(QueryResults.class);
+    String getQuery();
 
-    private static final Splitter SESSION_HEADER_SPLITTER = Splitter.on('=').limit(2).trimResults();
-    private static final String USER_AGENT_VALUE = StatementClient.class.getSimpleName() +
-            "/" +
-            firstNonNull(StatementClient.class.getPackage().getImplementationVersion(), "unknown");
+    TimeZoneKey getTimeZone();
 
-    private final OkHttpClient httpClient;
-    private final String query;
-    private final AtomicReference<QueryResults> currentResults = new AtomicReference<>();
-    private final AtomicReference<String> setCatalog = new AtomicReference<>();
-    private final AtomicReference<String> setSchema = new AtomicReference<>();
-    private final Map<String, String> setSessionProperties = new ConcurrentHashMap<>();
-    private final Set<String> resetSessionProperties = Sets.newConcurrentHashSet();
-    private final Map<String, String> addedPreparedStatements = new ConcurrentHashMap<>();
-    private final Set<String> deallocatedPreparedStatements = Sets.newConcurrentHashSet();
-    private final AtomicReference<String> startedTransactionId = new AtomicReference<>();
-    private final AtomicBoolean clearTransactionId = new AtomicBoolean();
-    private final TimeZoneKey timeZone;
-    private final Duration requestTimeoutNanos;
-    private final String user;
+    boolean isRunning();
 
-    private final AtomicReference<State> state = new AtomicReference<>(State.RUNNING);
+    boolean isClientAborted();
 
-    public StatementClient(OkHttpClient httpClient, ClientSession session, String query)
-    {
-        requireNonNull(httpClient, "httpClient is null");
-        requireNonNull(session, "session is null");
-        requireNonNull(query, "query is null");
+    boolean isClientError();
 
-        this.httpClient = httpClient;
-        this.timeZone = session.getTimeZone();
-        this.query = query;
-        this.requestTimeoutNanos = session.getClientRequestTimeout();
-        this.user = session.getUser();
+    boolean isFinished();
 
-        Request request = buildQueryRequest(session, query);
+    StatementStats getStats();
 
-        JsonResponse<QueryResults> response = JsonResponse.execute(QUERY_RESULTS_CODEC, httpClient, request);
-        if ((response.getStatusCode() != HTTP_OK) || !response.hasValue()) {
-            state.compareAndSet(State.RUNNING, State.CLIENT_ERROR);
-            throw requestFailedException("starting query", request, response);
-        }
+    QueryStatusInfo currentStatusInfo();
 
-        processResponse(response.getHeaders(), response.getValue());
-    }
+    QueryData currentData();
 
-    private Request buildQueryRequest(ClientSession session, String query)
-    {
-        HttpUrl url = HttpUrl.get(session.getServer());
-        if (url == null) {
-            throw new ClientException("Invalid server URL: " + session.getServer());
-        }
-        url = url.newBuilder().encodedPath("/v1/statement").build();
+    QueryStatusInfo finalStatusInfo();
 
-        Request.Builder builder = prepareRequest(url)
-                .post(RequestBody.create(MEDIA_TYPE_JSON, query));
+    Optional<String> getSetCatalog();
 
-        if (session.getSource() != null) {
-            builder.addHeader(PRESTO_SOURCE, session.getSource());
-        }
-        if (session.getClientTags() != null && !session.getClientTags().isEmpty()) {
-            builder.addHeader(PRESTO_CLIENT_TAGS, Joiner.on(",").join(session.getClientTags()));
-        }
-        if (session.getClientInfo() != null) {
-            builder.addHeader(PRESTO_CLIENT_INFO, session.getClientInfo());
-        }
-        if (session.getCatalog() != null) {
-            builder.addHeader(PRESTO_CATALOG, session.getCatalog());
-        }
-        if (session.getSchema() != null) {
-            builder.addHeader(PRESTO_SCHEMA, session.getSchema());
-        }
-        builder.addHeader(PRESTO_TIME_ZONE, session.getTimeZone().getId());
-        if (session.getLocale() != null) {
-            builder.addHeader(PRESTO_LANGUAGE, session.getLocale().toLanguageTag());
-        }
+    Optional<String> getSetSchema();
 
-        Map<String, String> property = session.getProperties();
-        for (Entry<String, String> entry : property.entrySet()) {
-            builder.addHeader(PRESTO_SESSION, entry.getKey() + "=" + entry.getValue());
-        }
+    Map<String, String> getSetSessionProperties();
 
-        Map<String, String> statements = session.getPreparedStatements();
-        for (Entry<String, String> entry : statements.entrySet()) {
-            builder.addHeader(PRESTO_PREPARED_STATEMENT, urlEncode(entry.getKey()) + "=" + urlEncode(entry.getValue()));
-        }
+    Set<String> getResetSessionProperties();
 
-        builder.addHeader(PRESTO_TRANSACTION_ID, session.getTransactionId() == null ? "NONE" : session.getTransactionId());
+    Map<String, String> getAddedPreparedStatements();
 
-        return builder.build();
-    }
-
-    public String getQuery()
-    {
-        return query;
-    }
-
-    public TimeZoneKey getTimeZone()
-    {
-        return timeZone;
-    }
-
-    public boolean isRunning()
-    {
-        return state.get() == State.RUNNING;
-    }
-
-    public boolean isClientAborted()
-    {
-        return state.get() == State.CLIENT_ABORTED;
-    }
-
-    public boolean isClientError()
-    {
-        return state.get() == State.CLIENT_ERROR;
-    }
-
-    public boolean isFinished()
-    {
-        return state.get() == State.FINISHED;
-    }
-
-    public StatementStats getStats()
-    {
-        return currentResults.get().getStats();
-    }
-
-    public QueryData currentData()
-    {
-        checkState(isRunning(), "current position is not valid (cursor past end)");
-        return currentResults.get();
-    }
-
-    public QueryStatusInfo currentStatusInfo()
-    {
-        checkState(isRunning(), "current position is not valid (cursor past end)");
-        return currentResults.get();
-    }
-
-    public QueryStatusInfo finalStatusInfo()
-    {
-        checkState(!isRunning(), "current position is still valid");
-        return currentResults.get();
-    }
-
-    public Optional<String> getSetCatalog()
-    {
-        return Optional.ofNullable(setCatalog.get());
-    }
-
-    public Optional<String> getSetSchema()
-    {
-        return Optional.ofNullable(setSchema.get());
-    }
-
-    public Map<String, String> getSetSessionProperties()
-    {
-        return ImmutableMap.copyOf(setSessionProperties);
-    }
-
-    public Set<String> getResetSessionProperties()
-    {
-        return ImmutableSet.copyOf(resetSessionProperties);
-    }
-
-    public Map<String, String> getAddedPreparedStatements()
-    {
-        return ImmutableMap.copyOf(addedPreparedStatements);
-    }
-
-    public Set<String> getDeallocatedPreparedStatements()
-    {
-        return ImmutableSet.copyOf(deallocatedPreparedStatements);
-    }
+    Set<String> getDeallocatedPreparedStatements();
 
     @Nullable
-    public String getStartedTransactionId()
-    {
-        return startedTransactionId.get();
-    }
+    String getStartedTransactionId();
 
-    public boolean isClearTransactionId()
-    {
-        return clearTransactionId.get();
-    }
+    boolean isClearTransactionId();
 
-    private Request.Builder prepareRequest(HttpUrl url)
-    {
-        return new Request.Builder()
-                .addHeader(PRESTO_USER, user)
-                .addHeader(USER_AGENT, USER_AGENT_VALUE)
-                .url(url);
-    }
+    boolean advance();
 
-    public boolean advance()
-    {
-        if (!isRunning()) {
-            return false;
-        }
-
-        URI nextUri = currentStatusInfo().getNextUri();
-        if (nextUri == null) {
-            state.compareAndSet(State.RUNNING, State.FINISHED);
-            return false;
-        }
-
-        Request request = prepareRequest(HttpUrl.get(nextUri)).build();
-
-        Exception cause = null;
-        long start = System.nanoTime();
-        long attempts = 0;
-
-        while (true) {
-            // back-off on retry
-            if (attempts > 0) {
-                try {
-                    MILLISECONDS.sleep(attempts * 100);
-                }
-                catch (InterruptedException e) {
-                    try {
-                        close();
-                    }
-                    finally {
-                        Thread.currentThread().interrupt();
-                    }
-                    state.compareAndSet(State.RUNNING, State.CLIENT_ERROR);
-                    throw new RuntimeException("StatementClient thread was interrupted");
-                }
-            }
-            attempts++;
-
-            JsonResponse<QueryResults> response;
-            try {
-                response = JsonResponse.execute(QUERY_RESULTS_CODEC, httpClient, request);
-            }
-            catch (RuntimeException e) {
-                cause = e;
-                continue;
-            }
-
-            if ((response.getStatusCode() == HTTP_OK) && response.hasValue()) {
-                processResponse(response.getHeaders(), response.getValue());
-                return true;
-            }
-
-            if (response.getStatusCode() != HTTP_UNAVAILABLE) {
-                state.compareAndSet(State.RUNNING, State.CLIENT_ERROR);
-                throw requestFailedException("fetching next", request, response);
-            }
-
-            if (Duration.nanosSince(start).compareTo(requestTimeoutNanos) > 0) {
-                state.compareAndSet(State.RUNNING, State.CLIENT_ERROR);
-                throw new RuntimeException("Error fetching next", cause);
-            }
-
-            if (isClientAborted()) {
-                return false;
-            }
-        }
-    }
-
-    private void processResponse(Headers headers, QueryResults results)
-    {
-        setCatalog.set(headers.get(PRESTO_SET_CATALOG));
-        setSchema.set(headers.get(PRESTO_SET_SCHEMA));
-
-        for (String setSession : headers.values(PRESTO_SET_SESSION)) {
-            List<String> keyValue = SESSION_HEADER_SPLITTER.splitToList(setSession);
-            if (keyValue.size() != 2) {
-                continue;
-            }
-            setSessionProperties.put(keyValue.get(0), keyValue.get(1));
-        }
-        resetSessionProperties.addAll(headers.values(PRESTO_CLEAR_SESSION));
-
-        for (String entry : headers.values(PRESTO_ADDED_PREPARE)) {
-            List<String> keyValue = SESSION_HEADER_SPLITTER.splitToList(entry);
-            if (keyValue.size() != 2) {
-                continue;
-            }
-            addedPreparedStatements.put(urlDecode(keyValue.get(0)), urlDecode(keyValue.get(1)));
-        }
-        for (String entry : headers.values(PRESTO_DEALLOCATED_PREPARE)) {
-            deallocatedPreparedStatements.add(urlDecode(entry));
-        }
-
-        String startedTransactionId = headers.get(PRESTO_STARTED_TRANSACTION_ID);
-        if (startedTransactionId != null) {
-            this.startedTransactionId.set(startedTransactionId);
-        }
-        if (headers.get(PRESTO_CLEAR_TRANSACTION_ID) != null) {
-            clearTransactionId.set(true);
-        }
-
-        currentResults.set(results);
-    }
-
-    private RuntimeException requestFailedException(String task, Request request, JsonResponse<QueryResults> response)
-    {
-        if (!response.hasValue()) {
-            if (response.getStatusCode() == HTTP_UNAUTHORIZED) {
-                return new ClientException("Authentication failed" +
-                        Optional.ofNullable(response.getStatusMessage())
-                                .map(message -> ": " + message)
-                                .orElse(""));
-            }
-            return new RuntimeException(
-                    format("Error %s at %s returned an invalid response: %s [Error: %s]", task, request.url(), response, response.getResponseBody()),
-                    response.getException());
-        }
-        return new RuntimeException(format("Error %s at %s returned HTTP %s", task, request.url(), response.getStatusCode()));
-    }
-
-    public void cancelLeafStage()
-    {
-        checkState(!isClientAborted(), "client is closed");
-
-        URI uri = currentStatusInfo().getPartialCancelUri();
-        if (uri != null) {
-            httpDelete(uri);
-        }
-    }
+    void cancelLeafStage();
 
     @Override
-    public void close()
-    {
-        // If the query is not done, abort the query.
-        if (state.compareAndSet(State.RUNNING, State.CLIENT_ABORTED)) {
-            URI uri = currentResults.get().getNextUri();
-            if (uri != null) {
-                httpDelete(uri);
-            }
-        }
-    }
-
-    private void httpDelete(URI uri)
-    {
-        Request request = prepareRequest(HttpUrl.get(uri))
-                .delete()
-                .build();
-        httpClient.newCall(request).enqueue(new NullCallback());
-    }
-
-    private static String urlEncode(String value)
-    {
-        try {
-            return URLEncoder.encode(value, "UTF-8");
-        }
-        catch (UnsupportedEncodingException e) {
-            throw new AssertionError(e);
-        }
-    }
-
-    private static String urlDecode(String value)
-    {
-        try {
-            return URLDecoder.decode(value, "UTF-8");
-        }
-        catch (UnsupportedEncodingException e) {
-            throw new AssertionError(e);
-        }
-    }
-
-    private enum State
-    {
-        /**
-         * submitted to server, not in terminal state (including planning, queued, running, etc)
-         */
-        RUNNING,
-        CLIENT_ERROR,
-        CLIENT_ABORTED,
-        /**
-         * finished on remote Presto server (including failed and successfully completed)
-         */
-        FINISHED,
-    }
+    void close();
 }
